Fix issue when no title for history item
Some Safari entries have no title. If this is the case, and they have
the same URL and datetime, then we can't sort them. Python doesn't allow
comparing strings and None.

This uses the solution here:
<https://stackoverflow.com/questions/60695318/sort-a-list-of-tuples-with-none-values-in-python3>

It does not affect the output. If the datetime were to ever be None,
this could cause an issue, potentially.